project/requirements.txt
google-api-python-client
yt-dlp
opencv-python
torch
transformers
pandas
tqdm
isodate
Pillow
numpy
future
scikit-learn
